TE-8082 Fixed bug with ProductStock bulk import on Postgres (#4877)
TE-8082 Postgres: Exception after the command: data:import:product-stock

+++ src/Pyz/Zed/DataImport/Business/Model/ProductStock/Writer/Sql/ProductStockSql.php
@@ -68,7 +68,7 @@ SELECT updated.idStock FROM updated UNION ALL SELECT inserted.id_stock FROM inse
            SELECT
              unnest(? :: VARCHAR []) AS sku,
              unnest(? :: VARCHAR []) AS stockName,
-             unnest(? :: INTEGER []) AS quantity,
+             unnest(? :: NUMERIC []) AS quantity,
              unnest(? :: BOOLEAN []) AS is_never_out_of_stock
          ) input
       INNER JOIN spy_stock on spy_stock.name = stockName
